Tests/LibWeb/Text/expected/Crypto/SubtleCrypto-ec-bad-import-key.txt

ECDSA P-256 PUB  - jwk   OK: DataError: Invalid key
ECDSA P-256 PUB  - spki  OK: DataError: Invalid key
ECDSA P-256 PRIV - jwk   OK: DataError: Invalid key
ECDSA P-256 PRIV - pkcs8 OK: DataError: Invalid key
ECDSA P-256 PRIV - pkcs8 OK: DataError: Invalid key
ECDSA P-384 PUB  - jwk   OK: DataError: Invalid key
ECDSA P-384 PUB  - spki  OK: DataError: Invalid key
ECDSA P-384 PRIV - jwk   OK: DataError: Invalid key
ECDSA P-384 PRIV - pkcs8 OK: DataError: Invalid key
ECDSA P-384 PRIV - pkcs8 OK: DataError: Invalid key
ECDSA P-521 PUB  - jwk   OK: DataError: Invalid key
ECDSA P-521 PUB  - spki  OK: DataError: Invalid key
ECDSA P-521 PRIV - jwk   OK: DataError: Invalid key
ECDSA P-521 PRIV - pkcs8 OK: DataError: Invalid key
ECDSA P-521 PRIV - pkcs8 OK: DataError: Invalid key
ECDH  P-256 PUB  - jwk   OK: DataError: Invalid key
ECDH  P-256 PUB  - spki  OK: DataError: Invalid key
ECDH  P-256 PRIV - jwk   OK: DataError: Invalid key
ECDH  P-256 PRIV - pkcs8 OK: DataError: Invalid key
ECDH  P-256 PRIV - pkcs8 OK: DataError: Invalid key
ECDH  P-384 PUB  - jwk   OK: DataError: Invalid key
ECDH  P-384 PUB  - spki  OK: DataError: Invalid key
ECDH  P-384 PRIV - jwk   OK: DataError: Invalid key
ECDH  P-384 PRIV - pkcs8 OK: DataError: Invalid key
ECDH  P-384 PRIV - pkcs8 OK: DataError: Invalid key
ECDH  P-521 PUB  - jwk   OK: DataError: Invalid key
ECDH  P-521 PUB  - spki  OK: DataError: Invalid key
ECDH  P-521 PRIV - jwk   OK: DataError: Invalid key
ECDH  P-521 PRIV - pkcs8 OK: DataError: Invalid key
ECDH  P-521 PRIV - pkcs8 OK: DataError: Invalid key
